some debug message appende to see why images are nto being downlaoded..
svn path=/trunk/KDE/kdeedu/marble/; revision=926976

+++ src/plugins/render/panoramio/jsonparser.h
@@ -33,8 +33,8 @@ struct panoramioDataStructure {
     QString photo_title; // Title of each photograph
     QString photo_url; // Url of each photograph
     QString photo_file_url;
-    qreal longitude;
-    qreal latitude;
+    int longitude;
+    int latitude;
     int width;
     int height;
     QString upload_date;
